save setting in chartMeasurements
