+++ frontend/src/metabase/visualizations/echarts/cartesian/chart-measurements/types.ts
@@ -1,3 +1,5 @@
+import type { ComputedVisualizationSettings } from "metabase/visualizations/types";
+
 export interface Padding {
   top: number;
   left: number;
@@ -28,4 +30,5 @@ export interface ChartMeasurements {
   bounds: ChartBoundsCoords;
   boundaryWidth: number;
   outerHeight: number;
+  axisEnabledSetting: ComputedVisualizationSettings["graph.x_axis.axis_enabled"];
 }
